@@ -1,8 +1,9 @@
 {
-  "author": "Van Carney <carney.van@gmail.com>",
-  "name": "wf-utils",
-  "description": "Small collection of utilities for Objects, Functions and Strings",
+  "author": "Van Schroeder <van@webfreshener.com>",
+  "name": "jsd-utils",
+  "description": "Object, Function and String Utilities",
   "keywords": [
+    "jsd",
     "object",
     "function",
     "string",
@@ -15,7 +16,7 @@
     "capitalize",
     "stripnull"
   ],
-  "version": "0.1.0",
+  "version": "1.0.0-beta.1",
   "preferGlobal": false,
   "main": "./index,js",
   "licenses": [
@@ -26,37 +27,29 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/webfreshener/wf-utils.git"
+    "url": "https://github.com/webfreshener/jsd-utils.git"
   },
   "dependencies": {},
   "devDependencies": {
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.18.2",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
     "babel-plugin-transform-remove-strict-mode": "0.0.2",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-register": "^6.18.0",
-    "chai": "^3.5.0",
-    "eslint": "^2.13.1",
-    "eslint-config-loopback": "^4.0.0",
-    "gulp": "^3.9.1",
-    "gulp-babel": "^6.1.2",
-    "gulp-browserify": "^0.5.1",
-    "gulp-clean": "^0.3.2",
-    "gulp-concat": "^2.6.1",
-    "gulp-documentation": "^3.0.0",
-    "gulp-insert": "^0.5.0",
-    "gulp-jshint": "^2.0.4",
-    "gulp-mocha": "^3.0.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-sourcemaps": "^2.2.0",
-    "gulp-webpack": "^1.5.0",
-    "jshint": "^2.9.4",
-    "mocha": "^3.2.0",
-    "nsp": "^2.1.0"
+    "babel-preset-env": "^1.6.1",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
+    "esdoc": "^1.0.4",
+    "esdoc-accessor-plugin": "^1.0.0",
+    "esdoc-node": "^1.0.3",
+    "esdoc-standard-plugin": "^1.0.0",
+    "eslint": "^4.16.0",
+    "jest": "^22.1.4",
+    "mocha": "^5.0.0",
+    "webpack": "^3.10.0"
   },
-  "homepage": "https://github.com/webfreshener/wf-utils.git",
+  "homepage": "https://github.com/webfreshener/jsd-utils.git",
   "scripts": {
-    "test": "gulp test"
+    "webpack": "webpack --config webpack.config.js",
+    "test": "jest",
+    "docs": "./node_modules/.bin/esdoc"
   }
 }
